cyberguard-ai/model_service/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
pydantic>=1.10
numpy>=1.24
tensorflow-cpu>=2.12
